# utils/indigo-service/doc/swagger.yaml
---
swagger: '2.0'
basePath: /v1
info:
  title: 'Indigo REST'
  version: '1.0'
produces:
  - application/json
tags:
  - name: library
    description: 'Methods to maniputate your libraries'
  - name: search
    description: 'Methods for search queries'
  - name: ketcher
    description: 'Methods for ketcher editor'
  - name: misc
    description: 'Other various methods'
paths:
  /libraries:
    get:
      description: 'Returns a library list'
      responses:
        200:
          description: 'A list of available libraries'
          schema:
            type: object
      tags:
        - library
    post:
      description: 'Creates a new library'
      consumes:
        - application/json
      parameters:
        - name: body
          in: body
          description: 'Library data payload (name and optional extra data)'
          required: true
      responses:
        201:
          description: 'Result object'
          schema:
            type: object
      tags:
        - library
  /libraries/{library_id}:
    get:
      description: 'Returns a library based on ID'
      parameters:
        - name: library_id
          in: path
          description: 'ID of a library to get'
          required: true
          type: string
      responses:
        200:
          description: 'Library object'
          schema:
            type: object
        404:
          description: 'Library not found'
      tags:
        - library
    put:
      description: 'Updates library data'
      consumes:
        - application/json
      parameters:
        - name: library_id
          in: path
          description: 'ID of a library to update'
          required: true
          type: string
        - name: body
          in: body
          description: 'New library data'
          required: true
      responses:
        200:
          description: 'Result object'
          schema:
            type: object
        404:
          description: 'Library not found'
      tags:
        - library
    delete:
      description: 'Removes a library'
      parameters:
        - name: library_id
          in: path
          description: 'ID of a library to delete'
          required: true
          type: string
      responses:
        200:
          description: 'Result object'
          schema:
            type: object
        404:
          description: 'Library not found'
      tags:
        - library
  /libraries/{library_id}/uploads:
    post:
      description: 'Uploads data to a selected library'
      consumes:
        - chemical/x-mdl-sdfile
        - application/x-gzip
        - application/gzip
      parameters:
        - name: library_id
          in: path
          description: 'ID of a target library'
          required: true
          type: string
        - name: body
          in: body
          description: 'Library contents payload'
          required: true
      responses:
        200:
          description: 'Result object'
          schema:
            type: object
        404:
          description: 'Library not found'
        415:
          description: 'Unsupported Media Type'
      tags:
        - library
  /libraries/{library_id}/uploads/{upload_id}:
    get:
      description: 'Check upload status for specified upload_id'
      parameters:
        - name: library_id
          in: path
          description: 'ID of a target library'
          required: true
          type: string
        - name: upload_id
          in: path
          description: 'ID of an upload task'
          required: true
          type: string
      responses:
        200:
          description: 'Result object'
          schema:
            type: object
      tags:
        - misc
  /render:
    post:
      description: 'Returns a molecule image'
      consumes:
        - application/json
      produces:
        - image/png
        - image/svg+xml
        - application/pdf
        - image/png;base64
      parameters:
        - name: body
          in: body
          description: 'A payload describing a structure and render options'
          required: true
      responses:
        200:
          description: 'A rendered molecule image'
          schema:
            type: file
        400:
            description: 'A problem with supplied client data'
      tags:
        - misc
  /info:
    get:
      description: 'Provides an information about service'
      responses:
        200:
          description: 'Information on various parts of the service'
          schema:
            type: object
      tags:
        - misc
  /search:
    post:
      description: 'Performs a search over specified libraries'
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: 'A payload describing search parameters'
          required: true
      responses:
        200:
          description: 'A list of found items'
          # TODO describe a response schema
          schema:
            type: object
      tags:
        - search
  /ketcher/knocknock:
    get:
      description: 'returns if ketcher availabl'
      produces:
        - text/plain
      tags:
        - ketcher  
  /ketcher/aromatize:
    post:
      description: 'Performs molecules aromatization'
      produces:
        - text/plain
      parameters:
        - name: body
          in: body
          description: 'Molecular structure'
          required: true
      responses:
        200:
          description: 'Aromatized structure'
          # TODO describe a response schema
          schema:
            type: object
      tags:
        - ketcher
  /ketcher/dearomatize:
    post:
      description: 'Performs molecules de-aromatization'
      produces:
        - text/plain
      parameters:
        - name: body
          in: body
          description: 'Molecular structure'
          required: true
      responses:
        200:
          description: 'De-aromatized structure'
          # TODO describe a response schema
          schema:
            type: object
      tags:
        - ketcher
  /ketcher/calculate_cip:
    post:
      description: 'Performs CIP descriptors calculation'
      produces:
        - text/plain
      parameters:
        - name: body
          in: body
          description: 'Molecular structure'
          required: true
      responses:
        200:
          description: 'Structure'
          # TODO describe a response schema
          schema:
            type: object
      tags:
        - ketcher